refactor structure of Search-related components
